refactor(Comments): rename component
follows a more consistent pattern where
the exported method matches the filename
